// TSIC-Core-Angular/src/frontend/tsic-app/src/app/core/api/models/models/RegisteredTeamDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type RegisteredTeamDto = {
    teamId: string;
    clubTeamId: number | string;
    clubTeamName: string;
    clubTeamGradYear: string;
    clubTeamLevelOfPlay: string;
    ageGroupId: string;
    ageGroupName: string;
    feeBase: number | string;
    feeProcessing: number | string;
    feeTotal: number | string;
    paidTotal: number | string;
    owedTotal: number | string;
};