// TSIC-Core-Angular/src/frontend/tsic-app/src/app/core/api/models/models/AgeGroupDto.ts
/* generated using openapi-typescript-codegen -- do not edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */
export type AgeGroupDto = {
    ageGroupId: string;
    ageGroupName: string;
    maxTeams: number | string;
    registeredCount: number | string;
    rosterFee: number | string;
    teamFee: number | string;
};